requirements.txt
Flask==1.1.1
Werkzeug==0.15.4
Keras==2.3.1
numpy==1.16.0
pandas==0.24.2
gunicorn
Pillow